requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
langchain~=0.3.1
pydantic~=2.9.2
flask~=3.0.3